requirements.txt
numpy
faiss-cpu
gspread
google-auth
sentence-transformers
google-generativeai
aiohttp
python-telegram-bot
flask